+++ design-system/analytics-next/src/cleanProps.ts
@@ -1,5 +1,0 @@
-export default function cleanProps(props: Record<string, any>) {
-  const { createAnalyticsEvent, ...cleanedProps } = props;
-
-  return cleanedProps;
-}