+++ design-system/analytics/src/index.js
@@ -1,5 +1,0 @@
-export { default as AnalyticsDecorator } from './AnalyticsDecorator';
-export { default as AnalyticsDelegate } from './AnalyticsDelegate';
-export { default as AnalyticsListener } from './AnalyticsListener';
-export { default as cleanProps } from './cleanProps';
-export { default as withAnalytics } from './withAnalytics';