+++ design-system/textfield/examples/00-basic.tsx
@@ -1,82 +1,29 @@
-import React, { Component } from 'react';
+import React from 'react';
+
+import { subtleHeading } from '@atlaskit/theme/colors';
+import { fontSize, gridSize } from '@atlaskit/theme/constants';
+import { headingSizes } from '@atlaskit/theme/typography';
 
 import Textfield from '../src';
 
-const eventResultStyle = {
-  borderStyle: 'dashed',
-  borderWidth: '1px',
-  borderColor: '#ccc',
-  padding: '0.5em',
-  color: '#ccc',
-  margin: '0.5em 0',
-};
-
-type State = { eventResult: string };
-
-export default class TextfieldExample extends Component<void, State> {
-  state = {
-    eventResult:
-      'Click into & out of the input above to trigger onBlur & onFocus.',
-  };
-
-  handleOnChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    this.setState({
-      eventResult: `onChange called with value: ${e.target.value}`,
-    });
-  };
-
-  handleOnBlur = () => {
-    this.setState({ eventResult: 'onBlur called' });
-  };
-
-  handleOnFocus = () => {
-    this.setState({ eventResult: 'onFocus called' });
-  };
-
-  render() {
-    const { eventResult } = this.state;
-
-    return (
-      <div>
-        <label htmlFor="event-handlers">Event Handlers</label>
-        <Textfield
-          name="event-handlers"
-          onChange={this.handleOnChange}
-          onBlur={this.handleOnBlur}
-          onFocus={this.handleOnFocus}
-        />
-        <div style={eventResultStyle}>{eventResult}</div>
-
-        <label htmlFor="default-value">
-          Default value (not the same as a placeholder)
-        </label>
-        <Textfield name="default-value" defaultValue="candy" />
-
-        <label htmlFor="disabled">Disabled</label>
-        <Textfield
-          name="disabled"
-          isDisabled
-          defaultValue="can't touch this..."
-        />
-
-        <label htmlFor="required">Required</label>
-        <Textfield name="required" isRequired />
-
-        <label htmlFor="invalid">Invalid</label>
-        <Textfield name="invalid" isInvalid />
-
-        <label htmlFor="placeholder">Placeholder</label>
-        <Textfield name="placeholder" placeholder="Click here to input..." />
-
-        <label htmlFor="auto-focus">Auto Focus</label>
-        <Textfield name="auto-focus" autoFocus />
-
-        <label htmlFor="spell-check">Spell Check</label>
-        <Textfield name="spell-check" spellCheck />
-
-        <label htmlFor="compact">Compact</label>
-        <Textfield name="compact" isCompact />
-      </div>
-    );
-  }
+export default function() {
+  return (
+    <div>
+      <label
+        htmlFor="basic"
+        style={{
+          fontSize: `${headingSizes.h200.size / fontSize()}em`,
+          fontStyle: 'inherit',
+          lineHeight: `${headingSizes.h200.lineHeight /
+            headingSizes.h200.size}`,
+          color: `${subtleHeading()}`,
+          fontWeight: 600,
+          marginTop: `${gridSize() * 2}px`,
+        }}
+      >
+        Basic text field
+      </label>
+      <Textfield name="basic" />
+    </div>
+  );
 }