+++ design-system/textfield/src/components/__tests__/visual-regression/textFieldSnapshotTest.ts
@@ -7,11 +7,11 @@ declare var global: any;
 
 // TODO - identify threshold or do not set cursor
 describe('Snapshot Test', () => {
-  it('Textfield basic should match production example', async () => {
+  it('Textfield variations should match production example', async () => {
     const url = getExampleUrl(
       'design-system',
       'textfield',
-      'basic',
+      'variations',
       global.__BASEURL__,
     );
     const image = await takeScreenShot(global.page, url);